Refactor | #47 | @lcomment | Redis repository 관련 리팩토링
