+++ cakk-api/src/main/java/com/cakk/api/service/user/SignService.java
@@ -16,7 +16,7 @@ import com.cakk.common.exception.CakkException;
 import com.cakk.domain.mysql.entity.user.User;
 import com.cakk.domain.mysql.repository.reader.UserReader;
 import com.cakk.domain.mysql.repository.writer.UserWriter;
-import com.cakk.domain.redis.repository.impl.TokenRedisRepository;
+import com.cakk.domain.redis.repository.TokenRedisRepository;
 
 @Service
 @RequiredArgsConstructor
