+++ cakk-domain/redis/src/main/java/com/cakk/domain/redis/annotation/RedisCustomTemplate.java
@@ -1,0 +1,26 @@
+package com.cakk.domain.redis.annotation;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.springframework.stereotype.Component;
+
+/**
+ * Indicates that an annotated class is a "RedisTemplate" (e.g. a data access object).
+ *
+ * <p>This annotation serves as a specialization of {@link Component @Component},
+ * allowing for implementation classes to be autodetected through classpath scanning.
+ *
+ * @author komment
+ * @see Component
+ */
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Component
+public @interface RedisCustomTemplate {
+}
